pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-socketio>=5.5.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "miniupnpc>=2.3.3",
    "psycopg2-binary>=2.9.10",
    "pystun>=0.1.0",
    "python-bitcoinlib>=0.12.2",
    "qrcode>=8.2",
    "cryptography>=45.0.5",
]
